+++ Sources/TStreamKit/Source/HTTPMediaSource.swift
@@ -123,8 +123,8 @@ final class HTTPMediaSource: MediaSource {
 
 // Forward the demuxer's output to the player.
 extension HTTPMediaSource: StreamDemuxerOutput {
-    func demuxerDidParseVideoFormat(_ codec: VideoCodec, extradata: Data?) {
-        delegate?.mediaSource(self, didParseVideoFormat: codec, extradata: extradata)
+    func demuxerDidParseVideoFormat(_ codec: VideoCodec, extradata: Data?, pixelAspect: PixelAspect?) {
+        delegate?.mediaSource(self, didParseVideoFormat: codec, extradata: extradata, pixelAspect: pixelAspect)
     }
     func demuxerDidProduceVideo(_ data: Data, codec: VideoCodec, pts: UInt64, dts: UInt64) {
         delegate?.mediaSource(self, didProduceVideo: data, codec: codec, pts: pts, dts: dts)